+++ raw_trainer/keep_prob.py
@@ -1,3 +1,3 @@
 import tensorflow as tf
 
-keep_prob = tf.placeholder(tf.float32)
+keep_prob = tf.placeholder(tf.float16)
